+++ src/index.js
@@ -1,25 +1,10 @@
 import './style.css';
-import addTaskScreen from './modules/module2.js';
-import Tasks from './modules/module3.js';
-
-let tasksArray = [];
+import addTaskScreen from './modules/addTask-function.js';
+import Tasks from './modules/obj-constructor.js';
+import { getTasks, addTaskStorage } from './modules/localStorage-functions.js';
 
 const enterNewTask = document.querySelector('.enter');
-
-const getTasks = () => {
-  if (localStorage.getItem('tasks') === null) {
-    tasksArray = [];
-  } else {
-    tasksArray = JSON.parse(localStorage.getItem('tasks'));
-  }
-  return tasksArray;
-};
-
-const addTaskStorage = (task) => {
-  tasksArray = getTasks();
-  tasksArray.push(task);
-  localStorage.setItem('tasks', JSON.stringify(tasksArray));
-};
+let tasksArray = [];
 
 const displayTasks = () => {
   tasksArray = getTasks();
@@ -28,8 +13,6 @@ const displayTasks = () => {
 
 enterNewTask.addEventListener('keypress', (e) => {
   if (e.keyCode === 13) {
-    // eslint-disable-next-line no-unused-expressions
-    e.preventDefault;
     tasksArray = getTasks();
     const task = new Tasks(enterNewTask.value, false, tasksArray.length + 1);
     addTaskStorage(task);
