+++ src/modules/clearAll-function.js
@@ -1,11 +1,15 @@
-const clearAllCompleted = (tasksArray) => {
-  tasksArray = JSON.parse(localStorage.getItem('tasks'));
+import { getTasks, addNewStorage } from './localStorage-functions.js';
+
+let tasksArray = [];
+
+const clearAllCompleted = () => {
+  tasksArray = getTasks();
   const newA = tasksArray.filter((e) => e.completed === false);
   tasksArray = newA;
   tasksArray.forEach((task, index) => {
     task.index = index + 1;
   });
-  localStorage.setItem('tasks', JSON.stringify(tasksArray));
+  addNewStorage(tasksArray);
 };
 
 export default clearAllCompleted;
