+++ src/style.css
@@ -40,7 +40,7 @@ body {
   color: #000;
 }
 
-#list ul {
+.line {
   border-top: 0.5px solid #d0d9d4;
   display: flex;
   padding: 1rem;
@@ -48,7 +48,7 @@ body {
   gap: 0.6rem;
 }
 
-#list ul.active {
+.line.active {
   background-color: #ffffbf;
 }
 
@@ -73,7 +73,7 @@ input[type="checkbox"]:checked + input {
   margin-left: auto;
 }
 
-#list ul input,
+.line input,
 h3,
 span {
   font-weight: 300;
